% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_computation.R
\name{analyse_pulsar_data}
\alias{analyse_pulsar_data}
\title{analyse_pulsar_data}
\usage{
analyse_pulsar_data(
  hormone_df,
  smoothing_fraction,
  g_values,
  extinction_threshold,
  peak_split_depth,
  sdr_coef,
  nearest_nadir_distance,
  sdr_assay = NULL,
  n_steps = const$n_steps_constant,
  y_axis_max = NULL
)
}
\arguments{
\item{hormone_df}{Pulsar format data frame as produced by read_pulsar_input_file
or read_pulsar_input_folder file. Column 1 = sample number; column 2 = time, columns
3 to n are individual data series.}

\item{smoothing_fraction}{Proportional width of the total time of session
to be used as the smoothing window for lowess smoothing. Should be between 0 and 1}

\item{g_values}{Five-item vector of double defining the cutoff function g(n)
for peak identification (cf. Merriam and Wachter, 1982)}

\item{extinction_threshold}{Minimum value allowed in data series. Non-negative
values less than extinction_threshold are replaced with extinction_threshold in
function clean_data}

\item{peak_split_depth}{Size of drop and subsequent rise (in rescaled concentration
units) within a first pass peak that causes the peak to be redefined as two overlapping
events. See fsm_split_peaks for computational details.}

\item{sdr_coef}{Three-item vector of double defining the coefficients in the equation
used to estimate assay variability for rescaling (see function gen_rescaled for details).
Terms should be in the order: quadratic, linear, constant.}

\item{nearest_nadir_distance}{Amplitude is computed as the difference between the
maximum raw concentration in peak/pulse sequence and the raw concentration of the
"nearest nadir". The nearest nadir is the minimum raw concentration among n sample
points before the max point in the peak/pulse event. The nearest_nadir_distance
parameter sets the value of n. This value will default to 3, a sensible distance for
LH data.}

\item{sdr_assay}{Constant estimate of assay variation. Not used in the current implementation}

\item{n_steps}{Controls iteration of lowess algorithm. Default value of 6 replicates modified
lowess of Merriam and Wachter, 1982.}

\item{y_axis_max}{If NULL (default) plot y-axis limits are set to the min/max of their
individual data series. If a value is provided, y-axis limits are set to (0, provided value).}
}
\value{
List of outputs from pulsar_one. Each containts fields plot, peak_descriptors,
peak_summary and animal_summary.
}
\description{
Top level Pulsar routine. Accepts prepared data frame, performs pulsar peak
location analysis on all data series, by passing to function pulsar_one. Returns
list of analysis output data structures.
}
