% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_utilities.R
\name{check_data_integrity}
\alias{check_data_integrity}
\title{check_data_integrity}
\usage{
check_data_integrity(hormone_df)
}
\arguments{
\item{hormone_df}{A df with Sample No., Time and Data columns as produced by read_pulsar_input_file
and read_pulsar_input_foder.}
}
\description{
Accepts a df with Sample No., Time and Data columns as produced by read_pulsar_input_file
and read_pulsar_input_foder. Checks for a variety of conditions that might indicate a problem
with the data and issues warnings.
}
