% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_utilities.R
\name{clean_data}
\alias{clean_data}
\title{clean_data}
\usage{
clean_data(raw_data_df, threshold)
}
\arguments{
\item{raw_data_df}{The df is expected to have a column Concentration as
produced by read_pulsar_input_file and read_pulsar_input_folder.}

\item{threshold}{A double defining the minimum value in the cleaned
data series. Represents the sensitivity limit of the assay. May be 0.}
}
\value{
Copy of original df with cleaned Concentration column
}
\description{
Accepts a df containing a single data series. Removes negative readings.
Replaces non-negative readings below extinction threshold with threshold value.
Returns the cleaned version.
}
