% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFeatures.R
\name{estrada.class}
\alias{estrada.class}
\title{Estrada class of a graph}
\usage{
estrada.class(Graph, evthresh = 0.001)
}
\arguments{
\item{Graph}{pxp adjacency matrix for a graph}

\item{evthresh}{tolerance for a zero eigenvalue}
}
\value{
Estrada index (1-4)
}
\description{
Estrada proposes that graphs can be classified in 1 of 4 classes. We call this the Estrada index.
These are:
  I. Expander
 II. Cluster
III  Core-Periphery
IV   Mixed
}
\references{
Estrada, E. (2007). Topological structural classes of complex networks. Physical Review E - Statistical, Nonlinear, and Soft Matter Physics, 75(1), 1–12. doi:10.1103/PhysRevE.75.016103
}

