% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_computation.R
\name{find_peaks}
\alias{find_peaks}
\title{find_peaks}
\usage{
find_peaks(rescaled_conc, g_values)
}
\arguments{
\item{rescaled_conc}{Vector of rescaled concentration values as returned from gen_rescaled
or equivalent.}

\item{g_values}{Numeric values of function g(n). Values are rescaled cutoffs for sequence of length
n. In this implementation, n ranges from 1 to 5, inclusive.}
}
\value{
Named list containing peak_flags and peak_score.
\itemize{
\item peak_flags: For each point in input vector, 0 if point is not in a
peak/pulse as defined by g(n) and 1 if point is in peak/pulse.
\item peak_score: For each point in input vector contains the length (number
of consecutive points) in the peak containing the point, or 0 if the point
is not identified as being in a peak.
}
}
\description{
Using the g_values for function g(n), a peak is any set of n
consecutive points whose rescaled values are greater than g(n). Function g(n) decreases
with increasing n, so you will identify narrower high peaks and wider low peaks.
}
