% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPulsarSelect.R
\name{findConfFile}
\alias{findConfFile}
\title{find config file}
\usage{
findConfFile(name = "")
}
\arguments{
\item{name}{name of default config or path to config file. Empty string to use batchtools to search in the file system.}
}
\description{
Find a config file from batchtools or default file from pulsar
}
\seealso{
findTemplateFile
}
