% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPulsarSelect.R
\name{findTemplateFile}
\alias{findTemplateFile}
\title{find template file}
\usage{
findTemplateFile(name)
}
\arguments{
\item{name}{name of default template or path to template file. Empty string to use batchtool to search in the file system.}
}
\description{
Find a config file from batchtools or default file from pulsar
}
\seealso{
findConfFile
}
