% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_computation.R
\name{fsm_split_peaks}
\alias{fsm_split_peaks}
\title{fsm_split_peaks}
\usage{
fsm_split_peaks(peak_points, peak_split_depth, rescaled_conc)
}
\arguments{
\item{peak_points}{Named list containing peak_flags and peak_scores as returned by function
find_peaks}

\item{peak_split_depth}{Double defining the size (down and up) of change in the rescaled concentration
values of a peak sequence that triggers separation into separate peaks.}

\item{rescaled_conc}{Vector of normalised distances from baseline, as returned by function gen_rescaled
or equivalent.}
}
\value{
Named list containing peak_flags and peak_score.
\itemize{
\item peak_flags: For each point in input vector, 0 if point is not in a
peak/pulse as defined by g(n) and 1 if point is in peak/pulse.
\item peak_score: For each point in input vector contains the length (number
of consecutive points) in the peak containing the point, or 0 if the point
is not identified as being in a peak.
}
}
\description{
Walks the identified peak points (from function find_peaks) looking for overlapping
peaks that have incorrectly been identified as members of a single sequence.
Following Merriam (1982) we decide this has occurred if, within a sequence,
the rescaled values drop by parameter peak_split_depth, and then rise by
peak_split_depth. The break is defined as the low point of that dip and return.
This implementation uses a four-state fsm rather than the linear algorithm of the original PULSAR
code.
}
