% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_computation.R
\name{gen_amplitude}
\alias{gen_amplitude}
\title{gen_amplitude}
\usage{
gen_amplitude(raw_concentrations, peak_features, nearest_nadir_distance)
}
\arguments{
\item{raw_concentrations}{Raw concentration values}

\item{peak_features}{A named list of first and max points for each identified peak/pulse as returned
by function gen_peak_features}

\item{nearest_nadir_distance}{Amplitude is computed as the difference between the
maximum raw concentration in peak/pulse sequence and the raw concentration of the
"nearest nadir". The nearest nadir is the minimum raw concentration among n sample
points before the first point contained in the peak/pulse event. The nearest_nadir_distance
parameter sets the value of n. This value will default to 3, a sensible distance for
LH data.}
}
\value{
Vector of amplitudes for each peak
}
\description{
Generates amplitude for each found peak. Amplitude is computed as the
difference between the pulse's maximum raw concentration and the value of
the "nearest nadir". The nearest nadir is the lowest concentration value in
the previous nearest_nadir_distance points.
Amplitude is max raw conc minus raw conc at nearest nadir neighbour
}
