% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_utilities.R
\name{gen_base_pulsar_plot}
\alias{gen_base_pulsar_plot}
\title{gen_base_pulsar_plot}
\usage{
gen_base_pulsar_plot(
  animal_df,
  animal_id,
  smoothed_concentrations,
  time_column_label,
  concentration_column_label,
  pulsar_font,
  y_axis_max = NULL
)
}
\arguments{
\item{animal_df}{Data frame containing three columns: Sample number (ordinal),
sample time (in minutes), and raw data values.}

\item{animal_id}{Unique string identifier for current data series. Used in
image title}

\item{smoothed_concentrations}{Vector containing smoothed (baseline) data values. Must
be the same length as the raw data column in animal_df. See method gen_smoothed}

\item{time_column_label}{Axis label for x axis}

\item{concentration_column_label}{Axis label for y axis}

\item{pulsar_font}{Font object to insure good cross-platform behaviour. See
file pulsar_constants.R.}

\item{y_axis_max}{Allows the user to fix the upper limit of the y-axis. If NULL
y-axis is adjusted to the range of the raw concentration data values.}
}
\description{
Generates a ggplot of raw concentrations and smoothed concentrations.
If there are identified peaks, those points can be added later as
additional ggplot layers (see summarise_pulsar_results).
}
