% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_computation.R
\name{gen_peak_descriptors}
\alias{gen_peak_descriptors}
\title{gen_peak_descriptors}
\usage{
gen_peak_descriptors(
  animal_id,
  times,
  raw_concentrations,
  smoothed_concentrations,
  peak_features,
  nearest_nadir_distance
)
}
\arguments{
\item{animal_id}{Unique string identifier for the data series}

\item{times}{Time each assay sample was recorded}

\item{raw_concentrations}{Raw concentration values}

\item{smoothed_concentrations}{The values of the corresponding baseline, Currently lowess smoothed as produced
by function gen_smoothed}

\item{peak_features}{A named list of first and max points for each identified peak/pulse as returned
by function gen_peak_features}

\item{nearest_nadir_distance}{Amplitude is computed as the difference between the
maximum raw concentration in peak/pulse sequence and the raw concentration of the
"nearest nadir". The nearest nadir is the minimum raw concentration among n sample
points before the first point contained in the peak/pulse event. The nearest_nadir_distance
parameter sets the value of n. This value will default to 3, a sensible distance for
LH data.}
}
\value{
Data frame containing peak descriptor values (columns) for each peak (rows)
}
\description{
Generates summary information for found peaks using raw concentrations, smoothed concentrations
and peak_features (first and maximum points) as produced by method gen_peak_features. Summary
values are start times, maximum concentration time, maximum concentration
value, peak amplitude, and peak length for each identified peak/pulse.
}
