% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_computation.R
\name{gen_peak_features}
\alias{gen_peak_features}
\title{gen_peak_features}
\usage{
gen_peak_features(peak_points, rescaled_conc)
}
\arguments{
\item{peak_points}{Named list containing peak_flags and peak_scores as returned by function
fsm_split_peaks and trim_end_peaks (if desired)}

\item{rescaled_conc}{Vector of normalised distances from baseline, as returned by function gen_rescaled
or equivalent.}
}
\value{
Data frame containing first point (FirstPoint), point of maximum
  concentration (MaxPoint) and length (Length) for each identified peak/pulse
  sequence. If no peaks are found, returns a df with 0 rows and 0 cols !!! TBD change to null?
}
\description{
Accepts a peak_points list ($flag and $score) and produces
a descriptor df (first, max and length). The flags should be generated
by running find_peaks and fsm_split_peaks. Points not in a peak are 0.
Other peak flags show in peak, overlap point or ambiguous point as defined
by the corresponding constants (see file puslar_constants.R)
}
