% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_computation.R
\name{gen_peak_plot}
\alias{gen_peak_plot}
\title{gen_peak_plot}
\usage{
gen_peak_plot(
  animal_id,
  animal_df,
  smoothed_concentrations,
  peak_points,
  peak_features,
  nearest_nadir_distance,
  y_axis_max = NULL
)
}
\arguments{
\item{animal_id}{String serving as unique identified for the current data series}

\item{animal_df}{A pulsar format data frame (Sample, Time, Concentration) containing the
raw data}

\item{smoothed_concentrations}{The values of the corresponding baseline (lowess smoothed) as produced
by function gen_smoothed}

\item{peak_points}{A named list containing peak_flags and peak_score defining the state of each
point in the data series, as produced by the peak locating functions find_peaks, fsm_split_peaks
and trim_end_peaks}

\item{peak_features}{A named list of first and max points for each identified peak/pulse as returned
by function gen_peak_features}

\item{nearest_nadir_distance}{Amplitude is computed as the difference between the
maximum raw concentration in peak/pulse sequence and the raw concentration of the
"nearest nadir". The nearest nadir is the minimum raw concentration among n sample
points before the max point in the peak/pulse event. The nearest_nadir_distance
parameter sets the value of n. This value will default to 3, a sensible distance for
LH data.}

\item{y_axis_max}{If NULL (default) plot y-axis limits are set to the min/max of their
individual data series. If a value is provided, y-axis limits are set to (0, provided value).}
}
\value{
A named list of the three output artefacts plot, peak_descriptors and
peak_summary
}
\description{
Accepts intermediate results of a pulsar analysis of a single animal data
series (id, raw data df, smoothed baseline values, peak flags and peak
features) and produces three standardised output artefacts describing the
identified peaks/pulses. These are a ggplot with peak points marked, a named
list containing descriptors (first point, maximum concentration, amplitude,
etc.) for each peak, and a named list comtaining summary descritpors (e.g.
average amplitude, average peak interval, peak frequency) across all
identified peaks.
}
