% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_computation.R
\name{gen_peak_summary}
\alias{gen_peak_summary}
\title{gen_peak_summary}
\usage{
gen_peak_summary(
  animal_id,
  times,
  raw_concentrations,
  smoothed_concentrations,
  peak_features,
  nearest_nadir_distance
)
}
\arguments{
\item{animal_id}{Unique string identifier for the data series}

\item{times}{Time each assay sample was recorded}

\item{raw_concentrations}{Raw concentration values}

\item{smoothed_concentrations}{The values of the corresponding baseline, Currently lowess smoothed as produced
by function gen_smoothed}

\item{peak_features}{A named list of first and max points for each identified peak/pulse as returned
by function gen_peak_features}

\item{nearest_nadir_distance}{Amplitude is computed as the difference between the
maximum raw concentration in peak/pulse sequence and the raw concentration of the
"nearest nadir". The nearest nadir is the minimum raw concentration among n sample
points before the first point contained in the peak/pulse event. The nearest_nadir_distance
parameter sets the value of n. This value will default to 3, a sensible distance for
LH data.}
}
\value{
Data frame containing summary values across all peaks.
}
\description{
Generates summary information across all peaks: average amplitude and standard deviation
of amplitudes, average peak interval and standard deviation of peak intervals, and peak
frequency (peaks per hour)
}
