% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_computation.R
\name{gen_rescaled}
\alias{gen_rescaled}
\title{gen_rescaled}
\usage{
gen_rescaled(
  raw_concentrations,
  smoothed_concentrations,
  sdr_coef,
  sdr_assay = NULL
)
}
\arguments{
\item{raw_concentrations}{Vector of raw concentration values for single animal.}

\item{smoothed_concentrations}{Vector of computed baseline values. In current implementation
these are the output of function gen_smoothed, but alternative baseline computations could
be employed.}

\item{sdr_coef}{Three-item vector of double defining the coefficients in the equation
used to estimate assay variability for rescaling (see function gen_rescaled for details).
Terms should be in the order: quadratic, linear, constant.}

\item{sdr_assay}{Constant estimate of assay variation. Not used in the current implementation}
}
\value{
Vector of nomrmalise values.
}
\description{
Normalises raw concentrations as (raw-smooth)/noise.
Noise can be either a fixed estimate (if sdr_assay is supplied; not used in this
implementation) or an interpolated value computed for each data point as ax^2 + bx + c, if
sdr_coef(a,b,c) is supplied. See Merriam and Wachter, 1982 for how to compute
of a, b and c from assay replicates.
}
