% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_computation.R
\name{gen_smoothed}
\alias{gen_smoothed}
\title{gen_smoothed}
\usage{
gen_smoothed(
  times,
  raw_concentrations,
  smoothing_fraction,
  n_steps = const$n_steps_constant
)
}
\arguments{
\item{times}{Vector of time values}

\item{raw_concentrations}{Vector of raw concentration values for single animal.}

\item{smoothing_fraction}{Proportional width of the total time of session
to be used as the smoothing window for lowess smoothing. Should be between 0 and 1}

\item{n_steps}{Controls iteration of lowess algorithm. Value of 6 replicates modified
lowess of Merriam and Wachter, 1982.}
}
\value{
Vector of smoothed (i.e. baseline) values.
}
\description{
Calls lowess on time x concentration to generate smoothing line. The lowess method is in
library stats
}
