% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_utilities.R
\name{get_current_time_string}
\alias{get_current_time_string}
\title{get_current_time_string}
\usage{
get_current_time_string()
}
\description{
Generates a string time stamp. Used for outfile naming.
}
