% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_computation.R
\name{get_near_nadir_conc}
\alias{get_near_nadir_conc}
\title{get_near_nadir_conc}
\usage{
get_near_nadir_conc(
  raw_concentrations,
  max_point_index,
  nearest_nadir_distance
)
}
\arguments{
\item{raw_concentrations}{Raw concentration values}

\item{max_point_index}{Integer value. Index of a point in the data vector for which the
nearest nadir concentration is to be calculated.}

\item{nearest_nadir_distance}{Amplitude is computed as the difference between the
maximum raw concentration in peak/pulse sequence and the raw concentration of the
"nearest nadir". The nearest nadir is the minimum raw concentration among n sample
points before the max point in the peak/pulse event. The nearest_nadir_distance
parameter sets the value of n. This value will default to 3, a sensible distance for
LH data.}
}
\value{
Vector of indices of the nearest nadir of each peak.
}
\description{
Given the index of a peak maximum, returns the raw concentration at the "nearest nadir" point.
This is the point with the lowest raw concentration in a window p samples prior to the maximum
concentration point of the peak. The window size p is a parameter of the algorithm passed
as argument nearest_nadir_distance. These values are used in the computation of amplitude.
}
