% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_computation.R
\name{get_near_nadir_indices}
\alias{get_near_nadir_indices}
\title{get_near_nadir_indices}
\usage{
get_near_nadir_indices(
  raw_concentrations,
  peak_features,
  nearest_nadir_distance
)
}
\arguments{
\item{raw_concentrations}{Raw concentration values}

\item{peak_features}{A named list of first and max points for each identified peak/pulse as returned
by function gen_peak_features}

\item{nearest_nadir_distance}{Amplitude is computed as the difference between the
maximum raw concentration in peak/pulse sequence and the raw concentration of the
"nearest nadir". The nearest nadir is the minimum raw concentration among n sample
points before the max point in the peak/pulse event. The nearest_nadir_distance
parameter sets the value of n. This value will default to 3, a sensible distance for
LH data.}
}
\value{
Vector of indices of the nearest nadir of each peak.
}
\description{
For each located peak/pulse, finds the index of the "nearest nadir" point. This is the
point with the lowest raw concentration in a window p samples prior to the maximum
concentration point of the peak. The window size p is a parameter of the algorithm passed
as argument nearest_nadir_distance. These values are used in the computation of amplitude.
}
