% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getEnvir}
\alias{getEnvir}
\title{Get calling environment}
\usage{
getEnvir(x)
}
\description{
Generic function for extracting a environment from an S3 object \code{x}. Designed as a generic getter for a stored calling environment of the original function.that produces the
}

