% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getLamPath}
\alias{getLamPath}
\title{Lambda path}
\usage{
getLamPath(max, min, len, log = FALSE)
}
\arguments{
\item{max}{numeric, maximum lambda value}

\item{min}{numeric, minimum lambda value}

\item{len}{numeric/int, length of lambda path}

\item{log}{logical, should the lambda path be log-spaced}
}
\value{
numeric vector of lambdas
}
\description{
Generate a lambda path sequence in descending order, equally or log spaced.
}

