% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getMaxCov}
\alias{getMaxCov}
\title{Max absolute value of cov matrix}
\usage{
getMaxCov(x, cov = isSymmetric(x), diag = FALSE)
}
\arguments{
\item{x}{A matrix/Matrix of data or covariance}

\item{cov}{Flag if \code{x} is a covariance matrix, Set False is \code{x} is an nxp data matrix. By default, if \code{x} is symmetric, assume it is a covariance matrix.}

\item{diag}{Flag to include diagonal entries in the max}
}
\description{
Get the maximum absolute value of a covariance matrix.
}

