% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFeatures.R
\name{graph.diss}
\alias{graph.diss}
\title{Graph dissimilarity}
\usage{
graph.diss(G, sim = FALSE, loops = FALSE)
}
\arguments{
\item{G}{adjacency matrix (dense or sparse) of a graph.}

\item{sim}{bool: return Graph similarity instead (which is just 1-dissimilarity)}

\item{loops}{bool: should self loops be considered}
}
\value{
a pxp dissimilarity matrix
}
\description{
Dissimilarity matrix of a graph is here defined as the number of shared neighbors any two nodes.
}
\references{
Bochkina, N. (2015). Selection of the Regularization Parameter in Graphical Models using a Priori Knowledge of Network Structure, arXiv: 1509.05326.
}

