% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{opt.index}
\alias{opt.index}
\alias{opt.index<-}
\title{Optimal index}
\usage{
opt.index(obj, criterion = "gcd")

opt.index(obj, criterion = names(value)) <- value
}
\arguments{
\item{obj}{a pulsar or batch.pulsar object}

\item{criterion}{a summary statistic criterion for lambda selection. If value is not named, default to gcd.}

\item{value}{Integer index for optimal lambda by criterion}
}
\description{
Get or set the optimal index of the lambda path, as determined by a given criterion. \code{value} must be a numeric/int.
}
\seealso{
\code{\link{get.opt.index}}
}
