% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.pulsar}
\alias{plot.batch.pulsar}
\alias{plot.pulsar}
\title{Plot a pulsar output}
\usage{
\method{plot}{pulsar}(x, ...)

\method{plot}{batch.pulsar}(x, ...)
}
\arguments{
\item{x}{a pulsar or batch.pulsar object}

\item{...}{system reserved usage (ignored)}

\item{scale}{Flag to scale non-StARS criterion to StARS range (or 1)}

\item{invlam}{Flag to plot 1/lambda}

\item{loglam}{Flag to plot log[lambda]}

\item{legends}{Flag to plot legends}
}
\description{
Plot a pulsar output
}
\details{
If both invlam and loglam are given, log[1/lambda] is plotted
}

