% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.pulsar}
\alias{print.batch.pulsar}
\alias{print.pulsar}
\title{Print S3 object \code{pulsar} and \code{batch.pulsar}}
\usage{
\method{print}{pulsar}(x, ...)

\method{print}{batch.pulsar}(x, ...)
}
\arguments{
\item{x}{a \code{pulsar} or \code{batch.pulsar} object}

\item{...}{Ignored}
}
\description{
Print information about the model, path length, graph dimension, criterion and optimal indices, if defined.
}

