% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.pulsar.refit}
\alias{print.pulsar.refit}
\title{Print S3 object \code{pulsar.refit}}
\usage{
\method{print}{pulsar.refit}(x, ...)
}
\arguments{
\item{x}{a \code{pulsar.refit}. output from \code{refit}}

\item{...}{ignored}
}
\description{
Print information about the model, path length, graph dimension, criterion and optimal indices and graph sparsity.
}

