% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar-package.R
\docType{package}
\name{pulsar-package}
\alias{pulsar-package}
\title{The pulsar package}
\description{
Graphical model selection with the pulsar package
}
\details{
This package provides methods to select a sparse, undirected graphical model by choosing a penalty parameter (lambda or \eqn{\lambda}) among many possible, ordered, values of lambda. We use an implementation of the Stability Approach to Regularization Selection (StARS, see references) loosely based on the \pkg{huge} package.

However, \pkg{pulsar} includes some major differences from other R packages for graphical model estimation and selection (\pkg{glasso}, \pkg{huge}, \pkg{QUIC}, \pkg{XMRF}, \pkg{clime}, \pkg{flare}, etc). The underlying graphical model is computed by passing a function as an argument to \code{\link{pulsar}}, so any algorithm for penalized graphical models can be used in this framework (see \code{\link{pulsar-function}} for more details), including those from the above packages. Therefore, \pkg{pulsar} bring computational experiments under one roof by separating the subsampling and calculation of summary criteria from the user-specified core model. Because of this, and unlike other packages, a typical workflow would perform subsampling first (via the \code{\link{pulsar}}) and then refit the model on the full dataset, using \code{\link{refit}}.

Edge stability (StARS) selects good models, but it can be inefficient for large graphs and when many subsamples are required. \code{\link{pulsar}} can compute upper and lower bounds for the StARS criterion after only \eqn{2} subsamples. This allows us to ditch lambda values that are far from StARS threshold, reducing computation time for the rest of the \eqn{N-2} subsamples.

Additionally we introduce additional subsampling-based summary criterion which can be used for model selection, for example to augment the choice of the StARS threshold. We have shown graphlet stability improves empirical performance over StARS but other criteria are also offered.

Subsampling amounts to running the specified core model for \eqn{N} indepedent computations. Using the \pkg{BatchJobs} framework, we provide a simple wrapper, \code{batch.pulsar}, for running \code{\link{pulsar}} in embarassingly parallel mode in an hpc environment. Summary criteria are computed using a Map/Reduce strategy, which lowers memory burden for large models.
}
\references{
Müller, C. L., Bonneau, R. A., & Kurtz, Z. D. (2016).Generalized Stability Approach for Regularized Graphical Models.arXiv: http://arxiv.org/abs/1605.07072.
}
\seealso{
\code{\link{pulsar-function}}, \code{\link{pulsar}}, \code{\link{batch.pulsar}}
}

