% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_utilities.R
\name{read_pulsar_input_file}
\alias{read_pulsar_input_file}
\title{read_pulsar_input_file}
\usage{
read_pulsar_input_file(pulsar_file_name)
}
\arguments{
\item{pulsar_file_name}{Complete file path to a correctly formatted input
file}
}
\value{
A named list (ExperimentName, HormoneDF) containing the experiment
  name and a data frame in correct format for further pulsaR processing. Returns
  NULL on I/O or data formatting error.
}
\description{
Read and prepare a single pulsaR input data file (see format requirements below)
}
\details{
Input file must be .csv.

File format is:

\itemize{
\item Line 1: Experiment name (any string) as the first element
\item Line 2: Column header line. Column headers can be any strings. Data
series column headers are treated as animal IDs, so should be unique.
\item Lines 3 to 500: Data values
}
Columns should be:
\itemize{
\item Column 1: Sample numbers (1 to n)
\item Column 2: Times (in minutes)
\item Columns 3 to n: Data values (typically hormone readings). Each data series (typically
animal) is in a single column.
}
All data series in a single file use the same time values.
}
