% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_utilities.R
\name{read_pulsar_input_folder}
\alias{read_pulsar_input_folder}
\title{read_pulsar_input_folder}
\usage{
read_pulsar_input_folder(input_folder_name)
}
\arguments{
\item{input_folder_name}{Complete file path to a correctly formatted input
file}
}
\value{
A named list (ExperimentName, HormoneDF) containing the experiment
  name and a data frame combining all files in correct format for further
  pulsaR processing. Returns NULL on I/O or data formatting error
}
\description{
Read and prepare multiple pulsaR input data files (each containing exactly
one data series; see format requirements below) contained in a single folder.
}
\details{
Input files must be .csv

File format is:

\itemize{
\item Line 1: Experiment name (any string) as the first element
\item Line 2: Column header line. Data series column header is treated as animal ID
so should be unique across all files in the folder.
\item Lines 3 to 500: Data values
}
Columns should be:
\itemize{
\item Column 1: Sample numbers (1 to n)
\item Column 2: Times (in minutes)
\item Column 3 Data values (typically hormone readings) for a single data series (typically
an animal)
}
All data files in the folder must have the same time values.
}
