% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit.R
\name{refit}
\alias{refit}
\title{Refit pulsar model}
\usage{
refit(obj, criterion)
}
\arguments{
\item{obj}{a \code{pulsar} or \code{batch.pulsar} object}

\item{criterion}{a vector of criteria for refitting on full data. An optimal index must be defined for each or a warning will result.}
}
\value{
an S3 \code{pulsar.refit} object with members:
\itemize{
  \item est: the raw output from the graphical model function (what was used to generate \code{obj}) but applied to the full dataset.
  \item refit: a named list of adjacency matrices, for each optimal criterion in \code{obj} or specified in the \code{criterion} argument
  \item fun: the original function for estimating the graphical model
}
}
\description{
Run the supplied graphical model function on the whole dataset and refit with the selected lambda(s)
}

