% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_utilities.R
\name{restore_param_list}
\alias{restore_param_list}
\title{restore_param_list}
\usage{
restore_param_list()
}
\value{
Named list containing the six adjustable algorithm parameters (smoothing fraction,
g(n) values, extinction threshold, peak split depth, assay variability estimation coefficients
and nearest nadir distance), and the two fixed algorithm parameters (constant assay variability
term and number of steps for lowess).
}
\description{
Loads file param_list.Rdata if present, creating object
 param_list. If the Rdata file is not present, default values (following Merriam 1982)
 are supplied. Later routines can parse param_list for the individual parameter terms.
}
