% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_main.R
\name{run_pulsar}
\alias{run_pulsar}
\title{run_pulsar}
\usage{
run_pulsar(
  input_file_name,
  param_list = NULL,
  make_outfiles = TRUE,
  zip_files = FALSE,
  y_axis_max = NULL
)
}
\arguments{
\item{input_file_name}{Complete file path to a correctly formatted input file (see details below)}

\item{param_list}{Named list of algorithm parameters as returned by function restore_parameters.
List can be built manually if desired, but field names must be as specified.

\itemize{
\item smoothing_fraction: Proportional width of the total time of session
to be used as the smoothing window for lowess smoothing. Should be between 0 and 1.
\item g_values: Five-item vector of double defining the cutoff function g(n)
for peak identification (cf. Merriam and Wachter, 1982).
\item extinction_threshold: Minimum value allowed in data series. Non-negative
values less than extinction_threshold are replaced with extinction_threshold in
function clean_data.
\item peak_split_depth: Size of drop and subsequent rise (in rescaled concentration
units) within a first pass peak that causes the peak to be redefined as two overlapping
events. See fsm_split_peaks for computational details.
\item sdr_coef: Three-item vector of double defining the coefficients in the equation
used to estimate assay variability for rescaling (see function gen_rescaled for details).
Terms should be in the order: quadratic, linear, constant.
\item nearest_nadir_distance: Amplitude is computed as the difference between the
maximum raw concentration in peak/pulse sequence and the raw concentration of the
"nearest nadir". The nearest nadir is the minimum raw concentration among n sample
points before the max point of the peak/pulse event. The nearest_nadir_distance
parameter sets the value of n. This value will default to 3, a sensible distance for
LH data.
\item sdr_assay: Constant assay variability estimate. Not used in this implementation; may
be null in the param_list.
\item n_steps: Lowess iteration term. Defaults to 6 if null in param_list. This value matches
numeric output of modified lowess as described by Merriam and Wachter, 1982.
}}

\item{make_outfiles}{If TRUE (the default) output files are written to the working directory.}

\item{zip_files}{If TRUE, output files are bundled into a single zip file.}

\item{y_axis_max}{If NULL (default) plot y-axis limits are set to the min/max of their
individual data series. If a value is provided, y-axis limits are set to (0, provided value).}
}
\value{
List of pulsar result objects as produced by method analyse_pulsar_data. Each contains
plot, peak descriptors, peak summary and animal summary for one data series.
}
\description{
Top-level user-facing pulsaR method.
}
\details{
Accepts a file in Pulsar Otago input format (see below). Runs
complete analysis. Returns list of pulsar results (plots, peak descriptors,
peak summaries and animal summaries). Generates outfiles if
make_outfiles = TRUE (default).


Input file must be .csv.

File format is:

\itemize{
\item Line 1: Experiment name (any string) as the first element
\item Line 2: Column header line. Column headers can be any strings. Data
series column headers are treated as animal IDs, so should be unique.
\item Lines 3 to 500: Data values.
}
Columns should be:

\itemize{
\item Column 1: Sample numbers (1 to n).
\item Column 2: Times (in minutes).
\item Columns 3 to n: Data values (typically hormone readings). Each data series (typically
animal) is in a single column.
}
All data series in a single file use the same time values.
}
