% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_utilities.R
\name{save_parameters}
\alias{save_parameters}
\title{save_parameters}
\usage{
save_parameters(
  smoothing_fraction,
  g_values,
  extinction_threshold,
  peak_split_depth,
  sdr_coef,
  nearest_nadir_distance
)
}
\arguments{
\item{smoothing_fraction}{Proportional width of the total time of session
to be used as the smoothing window for lowess smoothing. Should be between 0 and 1}

\item{g_values}{Five-item vector of double defining the cutoff function g(n)
for peak identification (cf. Merriam and Wachter, 1982)}

\item{extinction_threshold}{Minimum value allowed in data series. Non-negative
values less than extinction_threshold are replaced with extinction_threshold in
function clean_data}

\item{peak_split_depth}{Size of drop and subsequent rise (in rescaled concentration
units) within a first pass peak that causes the peak to be redefined as two overlapping
events. See fsm_split_peaks for computational details.}

\item{sdr_coef}{Three-item vector of double defining the coefficients in the equation
used to estimate assay variability for rescaling (see function gen_rescaled for details).
Terms should be in the order: quadratic, linear, constant.}

\item{nearest_nadir_distance}{Amplitude is computed as the difference between the
maximum raw concentration in peak/pulse sequence and the raw concentration of the
"nearest nadir". The nearest nadir is the minimum raw concentration among n sample
points before the max point of the peak/pulse event. The nearest_nadir_distance
parameter sets the value of n. This value will default to 3, a sensible distance for
LH data.}
}
\description{
Saves out a set of the 5 adjustable algorithm parameters as an Rdata
object for subsequent reload.
}
