% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_computation.R
\name{summarise_animal}
\alias{summarise_animal}
\title{summarise_animal}
\usage{
summarise_animal(animal_id, raw_concentrations)
}
\arguments{
\item{animal_id}{Unique string identifier for data series}

\item{raw_concentrations}{Vector of raw concentration values for single animal.}
}
\value{
Named list containing fields AnimalID, AvgConcentration and SdConcentration
}
\description{
Accepts vector of raw concentrations. Currently returns mean and stdev of raw
concentrations. Can be expanded in future as required.
}
