% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_computation.R
\name{trim_end_peaks}
\alias{trim_end_peaks}
\title{trim_end_peaks}
\usage{
trim_end_peaks(split_peak_points, rescaled_conc)
}
\arguments{
\item{split_peak_points}{Named list containing peak_flags and peak_scores as returned by function
fsm_split_peaks}

\item{rescaled_conc}{Vector of normalised distances from baseline, as returned by function gen_rescaled
or equivalent.}
}
\value{
Named list containing peak_flags and peak_score.
\itemize{
\item peak_flags: For each point in input vector, 0 if point is not in a
peak/pulse as defined by g(n), 1 if point is in peak/pulse and 2 if the point
was originally in a peak, but the maximum of that peak cannot be unambiguously
identified
\item peak_score: For each point in input vector contains the length (number
of consecutive points) in the peak containing the point, or 0 if the point
is not identified as being in a peak.
}
}
\description{
Ambiguous end points and their neighbour peaks are marked following Hackwell
rule: Is the next neighbour larger, and also still far enough from baseline
to qualify as a peak? If so, the maximum of this peak group is unambiguous,
and no change is needed. If not, every point in the peak needs to be flagged
as ambiguous. Ambiguous peak points are coloured differently and excluded
from the peak summary computation (see summarise_pulsar_result).
}
