% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{update.pulsar}
\alias{update.pulsar}
\title{Update a pulsar call}
\usage{
\method{update}{pulsar}(object, ..., evaluate = TRUE)
}
\arguments{
\item{object}{a n existing pulsar or batch.pulsar object}

\item{...}{arguments to \code{pulsar} to update}

\item{evaluate}{Flag to evaluate the function. If \code{FALSE}, the updated call is returned without evaluation}
}
\value{
If \code{evaluate = TRUE}, the fitted object - the same output as \code{pulsar} or \code{batch.pulsar}. Otherwise, the updated call.
}
\description{
Update a pulsar model with new or altered arguments. It does this by extracting the call stored in the object, updating the call and (by default) evaluating it in the environment of the original \code{pulsar} call.
}
\seealso{
\code{eval}, \code{\link{update}}, \code{\link{pulsar}}, \code{\link{batch.pulsar}}
}

