% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPulsarSelect.R
\name{batch.pulsar}
\alias{batch.pulsar}
\title{pulsar: batch mode}
\usage{
batch.pulsar(data, fun = huge::huge, fargs = list(),
  criterion = c("stars"), thresh = 0.1, subsample.ratio = NULL,
  lb.stars = FALSE, ub.stars = FALSE, rep.num = 20, seed = NULL,
  wkdir = getwd(), regdir = NA, init = "init", conffile = "",
  job.res = list(), cleanup = FALSE)
}
\arguments{
\item{data}{A \eqn{n*p} matrix of data matrix input to solve for the \eqn{p*p} graphical model}

\item{fun}{pass in a function that returns a list representing \eqn{p*p} sparse, undirected graphical models along the desired regularization path. The expected inputs to this function are: a data matrix input and a sequence of decreasing lambdas and must return a list or S3 object with a member \emph{named} \code{path}. This should be a list of adjacency matrices for each value of \code{lambda}. See \code{\link{pulsar-function}} for more information.}

\item{fargs}{arguments to argument \code{fun}. Must be a named list and requires at least one member \code{lambda}, a numeric vector with values for the penalty parameter.}

\item{criterion}{A character vector of selection statistics. Multiple criteria can be supplied. Only StARS can be used to automatically select an optimal index for the lambda path. See details for additional statistics.}

\item{thresh}{threshold (referred to as scalar \eqn{\beta} in StARS publication) for selection criterion. Only implemented for StARS. \code{thresh=0.1} is recommended.}

\item{subsample.ratio}{determine the size of the subsamples (referred to as \eqn{b(n)/n}). Default is 10*sqrt(n)/n for n > 144 or 0.8 otherwise. Should be strictly less than 1.}

\item{lb.stars}{Should the lower bound be computed after the first \eqn{N=2} subsamples (should result in considerable speedup and only implemented if stars is selected). If this option is selected, other summary metrics will only be applied to the smaller lambda path.}

\item{ub.stars}{Should the upper bound be computed after the first \eqn{N=2} subsamples (should result in considerable speedup and only implemented if stars is selected). If this option is selected, other summary metrics will only be applied to the smaller lambda path. This option is ignored if the lb.stars flag is FALSE.}

\item{rep.num}{number of random subsamples \eqn{N} to take for graph re-estimation. Default is \eqn{N=20}, but more is recommended for non-StARS criteria or if using edge frequencies as confidence scores.}

\item{seed}{A numeric seed to force predictable subsampling. Default is NULL. Use for testing purposes only.}

\item{wkdir}{set the working directory if different than \code{\link{getwd}}}

\item{regdir}{directory to store intermediate batch job files. Default will be a tempory directory}

\item{init}{text string appended to basename of the regdir path to store the batch jobs for the initial StARS variability estimate (ignored if `regdir` is NA)}

\item{conffile}{path or string to \code{\link[batchtools]{batchtools}} identify configuration file. See details for an explanation.}

\item{job.res}{named list of resources needed for each job (e.g. for PBS submission script). The format and members depends on configuration and template. See examples section for a Torque example}

\item{cleanup}{Flag for removing BatchJob registry files. Recommended FALSE unless you're sure intermediate data shouldn't be saved.}
}
\value{
an S3 object of class \code{pulsar} with a named member for each stability metric run. Within each of these are:
\itemize{
   \item summary: the summary statistic over \code{rep.num} graphs at each value of lambda
   \item criterion: the stability criterion used
   \item merge: the raw statistic over the \code{rep.num} graphs, prior to summarization
   \item opt.ind: index (along the path) of optimal lambda selected by the criterion at the desired threshold. Will return \eqn{0} if no optimum is found or \code{NULL} if selection for the criterion is not implemented.
  }
If \code{stars} is included as a criterion then additional arguments include
\itemize{
   \item lb.index: the lambda index of the lower bound at \eqn{N=2} samples if \code{lb.stars} flag is set to TRUE
   \item ub.index: the lambda index of the upper bound at \eqn{N=2} samples if \code{ub.stars} flag is set to TRUE
}

reg: Registry object. See \code{batchtools::makeRegistry}

id: Identifier for mapping graph estimation function. See \code{batchtools::batchMap}

call: the original function call
}
\description{
Run pulsar using stability selection, or another criteria, to select an undirected graphical model over a lambda-path.
}
\details{
The options for \code{criterion} statistics are:
\itemize{
   \item stars (Stability approach to regularization selection)
   \item gcd   (Graphlet correlation distance, requires the \pkg{orca} package)
   \item estrada (estrada class) see \code{\link{estrada.class}}
   \item sufficiency (Tandon & Ravikumar's sufficiency statistic)
}

If \code{conffile} is not given as a path to a file, we implement an extension of the \code{conf.file} option in \code{batchtools::makeRegistry}.
A file lookup is executed in the following order, and the first file found will be used
\itemize{\item File "batchtools.conf.R" in the path specified by the environment variable "R_BATCHTOOLS_SEARCH_PATH".
        \item File "batchtools.conf.R" in the current working directory.
        \item File "config.R" in the OS dependent user configuration directory as reported by via `rappdirs::user_config_dir("batchtools", expand = FALSE)` (e.g., on linux this usually resolves to "~/.config/batchtools/config.R").
        \item ".batchtools.conf.R" in the home directory ("~").
        \item The default file .batchtools.conf.R, located in the pulsar package directory and under the config subdirectory. This will execute pulsar in interactive/serial mode.
        If \code{conffile} is a string, get the default file batchtools.conf.{conffile}.R under pulsar's config subdirectory. Valid values are ['parallel', 'snow', 'torque'].
   }
For distributed computing environments with a queueing system, a simple default template for submitting the job (i.e. PBS script) is provided under the templates subdirectory. It is strongly recommended that a custom configfile and template be constructed to suit your cluster.
}
\examples{
\dontrun{
## Generate the data with huge:
library(huge)
set.seed(10010)
p <- 400 ; n <- 1200
dat   <- huge.generator(n, p, "hub", verbose=FALSE, v=.1, u=.3)
lams  <- getLamPath(.2, .01, len=40)
hugeargs  <- list(lambda=lams, verbose=FALSE)

## Run batch.pulsar using snow on 5 cores, and show progress.
options(mc.cores=5)
options(batchtools.progress=TRUE, batchtools.verbose=FALSE)
out <- batch.pulsar(dat$data, fun=huge::huge, fargs=hugeargs,
                 rep.num=20, criterion='stars', conffile='snow')
## Run batch.pulsar on a Torque cluster
## Give each job 1gb of memory and a limit of 30 minutes
resources <- list(memory="1GB", nodes="1", walltime="00:30:00")
out.p <- batch.pulsar(dat$data, fun=huge::huge, fargs=hugeargs,
                 rep.num=100, criterion=c('stars', 'gcd'), conffile='torque'
                 job.res=resources, regdir=file.path(getwd(), "testtorq"))
plot(out.p)
## check out the default torque conffile and template
pulsar::findConfFile('torque')
pulsar::findTemplateFile('simpletorque')
}
}
\references{
Müller, C. L., Bonneau, R., & Kurtz, Z. (2016). Generalized Stability Approach for Regularized Graphical Models. arXiv https://arxiv.org/abs/1605.07072

Liu, H., Roeder, K., & Wasserman, L. (2010). Stability approach to regularization selection (stars) for high dimensional graphical models. Proceedings of the Twenty-Third Annual Conference on Neural Information Processing Systems (NIPS).

Zhao, T., Liu, H., Roeder, K., Lafferty, J., & Wasserman, L. (2012). The huge Package for High-dimensional Undirected Graph Estimation in R. The Journal of Machine Learning Research, 13, 1059–1062.

Michel Lang, Bernd Bischl, Dirk Surmann (2017). batchtools: Tools for R to work on batch systems. The Journal of Open Source Software, 2(10). URL https://doi.org/10.21105/joss.00135.
}
\seealso{
\code{\link{pulsar}} \code{\link{refit}}
}
