% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFeatures.R
\name{estrada.class}
\alias{estrada.class}
\title{Estrada class}
\usage{
estrada.class(G, evthresh = 0.001)
}
\arguments{
\item{G}{a \eqn{p*p} adjacency matrix of a Graph}

\item{evthresh}{tolerance for a zero eigenvalue}
}
\value{
Estrada class (\eqn{1-4})
}
\description{
Estrada proposes that graphs can be classified into four different classes. We call this the Estrada class.
These are:
  I. Expander-like
 II. Cluster
III. Core-Periphery
IV.  Mixed.
}
\references{
Estrada, E. (2007). Topological structural classes of complex networks. Physical Review E - Statistical, Nonlinear, and Soft Matter Physics, 75(1), 1-12. doi:10.1103/PhysRevE.75.016103
}
