% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPulsarSelect.R
\name{findTemplateFile}
\alias{findTemplateFile}
\title{find template file}
\usage{
findTemplateFile(name)
}
\arguments{
\item{name}{name of default template or path to template file. Empty string to use batchtool to search in the file system.}
}
\description{
Find a config file from batchtools or default file from pulsar
}
\details{
Similar to the recipe used for template lookup in \code{batchtools::makeClusterFunctionsTORQUE}, \code{batchtools::makeClusterFunctionsSGE}, etc.

Either a path to a ‘brew’ template file (with extension
         "tmpl"), or a short descriptive name enabling the following
         heuristic for the file lookup:

   \enumerate{
     \item "batchtools.[name].tmpl" in the path specified by the
             environment variable "R_BATCHTOOLS_SEARCH_PATH".

     \item "batchtools.[name].tmpl" in the current working
             directory.

     \item "[name].tmpl" in the user config directory (see
             ‘user_config_dir’); on linux this is usually
             "~/.config/batchtools/[template].tmpl".

      \item ".batchtools.[name].tmpl" in the home directory.

     \item "[name].tmpl" in the batchtools installation directory
             in the subfolder "templates".

     \item "[name].tmpl" in the pulsar installation directory
             in the subfolder "templates".
   }
}
\seealso{
findConfFile
}
