% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools_config_template.R
\name{findTemplateFile}
\alias{findTemplateFile}
\title{find template file}
\usage{
findTemplateFile(name)
}
\arguments{
\item{name}{name of default template or path to template file.}
}
\description{
Find a config file from batchtools or default file from pulsar
}
\details{
See the batchtools functions \code{batchtools::findTemplateFile}, \code{batchtools::makeClusterFunctionsTORQUE}, \code{batchtools::makeClusterFunctionsSGE}, etc, to employ batchtools' default lookup scheme for template files. Supply the output of this function to the \code{template} argument to override batchtools' default.

In this case we look for "[name].tmpl" in the pulsar installation directory in the subfolder "templates".
}
\examples{
 \dontrun{
 cluster.functions = batchtools::makeClusterFunctionsTORQUE(
                     template=pulsar::findTemplateFile('simpletorque'))
 }
}
\seealso{
findConfFile
}
