% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getEnvir}
\alias{getEnvir}
\alias{getEnvir.default}
\title{Get calling environment}
\usage{
getEnvir(x)

\method{getEnvir}{default}(x)
}
\arguments{
\item{x}{S3 object to extract the environment}
}
\description{
Generic S3 method for extracting an environment from an S3 object. A getter for an explicitly stored environment from an S3 object or list... probably the environment where the original function that created the object was called from. The default method is a wrapper for \code{x$envir}.
}
\seealso{
\code{getCall}, \code{environment}, \code{parent.env}, \code{eval}
}
