% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.pulsar}
\alias{plot.pulsar}
\title{Plot a \code{pulsar} S3 object}
\usage{
\method{plot}{pulsar}(x, scale = TRUE, invlam = FALSE, loglam = FALSE, legends = TRUE, ...)
}
\arguments{
\item{x}{a \code{pulsar} or \code{batch.pulsar} object}

\item{scale}{Flag to scale non-StARS criterion to max StARS value (or 1)}

\item{invlam}{Flag to plot 1/lambda}

\item{loglam}{Flag to plot log[lambda]}

\item{legends}{Flag to plot legends}

\item{...}{ignored}
}
\description{
Plot a \code{pulsar} S3 object
}
\details{
If both invlam and loglam are given, log[1/lambda] is plotted
}
