% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.pulsar}
\alias{print.pulsar}
\alias{print.batch.pulsar}
\title{Print a \code{pulsar} and \code{batch.pulsar} S3 object}
\usage{
\method{print}{pulsar}(x, ...)

\method{print}{batch.pulsar}(x, ...)
}
\arguments{
\item{x}{a fitted \code{pulsar} or \code{batch.pulsar} object}

\item{...}{ignored}
}
\description{
Print information about the model, path length, graph dimension, criterion and optimal indices, if defined.
}
