% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{update.pulsar}
\alias{update.batch.pulsar}
\alias{update.pulsar}
\title{Update a pulsar call}
\usage{
update.pulsar(object, ..., evaluate = TRUE)

update.batch.pulsar(object, ..., evaluate = TRUE)
}
\arguments{
\item{object}{a pulsar or batch.pulsar object}

\item{...}{arguments to \code{pulsar} to update}

\item{evaluate}{Flag to evaluate the function. If FALSE, the updated call is returned without evaluation}
}
\value{
the same output as \code{pulsar} or \code{batch.pulsar}
}
\description{
Update and re-fit a model, with new or altered arguments. It does this by extracting the call stored in the object, updating the call and (by default) evaluating it in the environment of the original \code{pulsar} call.
}

